// fabrication/skin/src/api/gen/models/V1SessionTemplate.ts
export interface V1SessionTemplate {
    /**
     * 
     * @type {string}
     * @memberof V1SessionTemplate
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof V1SessionTemplate
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof V1SessionTemplate
     */
    command?: string;
    /**
     * 
     * @type {string}
     * @memberof V1SessionTemplate
     */
    restoreCommand?: string;
    /**
     * 
     * @type {string}
     * @memberof V1SessionTemplate
     */
    sessionFilePattern?: string;
    /**
     * 
     * @type {string}
     * @memberof V1SessionTemplate
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof V1SessionTemplate
     */
    icon?: string;
    /**
     * 
     * @type {boolean}
     * @memberof V1SessionTemplate
     */
    builtin?: boolean;
    /**
     * 
     * @type {V1ConfigLayer}
     * @memberof V1SessionTemplate
     */
    defaults?: V1ConfigLayer;
    /**
     * 
     * @type {V1SessionSchema}
     * @memberof V1SessionTemplate
     */
    sessionSchema?: V1SessionSchema;
}